Extract component frequency curve picker
